+++ x-core/lib/x/core/stream_parser.rb
@@ -26,6 +26,7 @@ module X
     # @yield [Object] each decoded JSON document from the stream
     # @return [void]
     # @raise [HTTPError] if the response is not successful
+    # @raise [InvalidResponse] if a line of the stream is not JSON
     # @example Process a streaming response
     #   handler.process(response: response, response_parser: parser) { |json| puts json }
     def process(response:, response_parser:, array_class: nil, object_class: nil, client: nil, on_body: nil, &block)
@@ -33,6 +34,8 @@ module X
       decode = lambda do |line|
         on_body&.call(line)
         response_parser.decode(line, array_class:, object_class:, client:)
+      rescue JSON::ParserError
+        raise InvalidResponse.new(response:)
       end
       read_lines(response:, decode:, &block)
     end